Add files via upload
Fixed problem in Iterative_BST which would incorrectly find-next of the largest item in the list and incorrectly find-prev of the smallest item in the list.
Fixed IterBSTester and RecBSTester to correctly test find-prev.
Added additional tests in IterBSTester and RecBSTester to more thoroughly test find-next and find-prev.

+++ src/question2/Sort_It.java
@@ -26,7 +26,7 @@ public class Sort_It {
 		if (current.leftChild != null)
 			index = inOrder(current.leftChild, retArr, index);
 		
-		retArr[index++] = current.value;
+		retArr[index++] = (int) current.value;
 		
 		if (current.rightChild != null)
 			index = inOrder(current.rightChild, retArr, index);
